Workflow: tweak test-local-subgraph (#2013)
**Problem**

At some point, the sdk-core/subgraph test cases started to fail with "socket hang up" error when querying subgraph.

After some research, it was discovered that it is due to the usage of an older version of the "cross-fetch" polyfill which didn't support node 20+, see [here](https://github.com/jasonkuhrt/graphql-request/issues/628) and [here](https://github.com/jasonkuhrt/graphql-request/pull/597).

**Solution**

The newer cross-fetch project solved this. The solution is to fix the "resolution" of the cross-fetch in this project to the latest, for now.

Additionally, this PR also:

a) include some minor refactoring around the test setup process,
b) test against both node 18 and 20 for SDK core, and test against node 20 for subgraph.